+++ src/main/java/io/renren/modules/sys/controller/SysUserController.java
@@ -101,7 +101,7 @@ public class SysUserController extends AbstractController {
 	 */
 	@GetMapping("/info/{userId}")
 	@RequiresPermissions("sys:user:info")
-	public CommonResult<Map<String, Object>> info(@PathVariable("userId") Long userId){
+	public CommonResult<Map<String, Object>> info(@PathVariable("userId") Integer userId){
 		SysUserEntity user = sysUserService.getById(userId);
 		
 		//获取用户所属的角色列表
